攻撃の判定と攻撃後の吹き飛ばしを追加 攻撃時のアニメーションが決まっていないから現状はその場で攻撃をして吹き飛ばされる

+++ Engine/GameObject/User/Enemy/Boss.h
@@ -50,7 +50,7 @@ private: // メンバ関数
 
 private:
 	//体力
-	int hitPoint_;
+	float hitPoint_;
 
 
 };
